Extract Weave component and shared turns module from App.svelte
