+++ qntx-plugins/loom/frontend/src/App.svelte
@@ -4,8 +4,10 @@
   import SessionList from './SessionList.svelte'
   import BranchBar from './BranchBar.svelte'
   import ClusterBar from './ClusterBar.svelte'
+  import Weave from './Weave.svelte'
   import Warp from './Warp.svelte'
   import { timeSpacers } from './timespacers'
+  import { parseTurns, turnKey, fmtTime, type Turn } from './turns'
 
   // --- Types ---
 
@@ -20,15 +22,6 @@
     paths?: Record<string, string>
   }
 
-  interface Turn {
-    speaker: string
-    text: string
-    weaveId: string
-    index: number
-    timestamp: number
-    branch: string
-    fullPath?: string
-  }
 
   interface Session {
     context: string
@@ -334,124 +327,8 @@
     }
   }
 
-  // --- Minimal markdown rendering (string methods only, no regex) ---
-
-  function escapeHtml(s: string): string {
-    return s.replaceAll('&', '&amp;').replaceAll('<', '&lt;').replaceAll('>', '&gt;')
-  }
-
-  function renderText(text: string): string {
-    const lines = text.split('\n')
-    let out = ''
-    let inCode = false
-    for (const line of lines) {
-      if (line.startsWith('```')) {
-        if (inCode) {
-          out += '</code></pre>'
-          inCode = false
-        } else {
-          inCode = true
-          out += '<pre class="dw-code"><code>'
-        }
-        continue
-      }
-      if (inCode) {
-        out += escapeHtml(line) + '\n'
-        continue
-      }
-      // Bold: **text**
-      let rendered = escapeHtml(line)
-      let result = ''
-      let pos = 0
-      while (pos < rendered.length) {
-        const start = rendered.indexOf('**', pos)
-        if (start < 0) { result += rendered.substring(pos); break }
-        const end = rendered.indexOf('**', start + 2)
-        if (end < 0) { result += rendered.substring(pos); break }
-        result += rendered.substring(pos, start) + '<b>' + rendered.substring(start + 2, end) + '</b>'
-        pos = end + 2
-      }
-      // Inline code: `code`
-      let final = ''
-      pos = 0
-      while (pos < result.length) {
-        const start = result.indexOf('`', pos)
-        if (start < 0) { final += result.substring(pos); break }
-        const end = result.indexOf('`', start + 1)
-        if (end < 0) { final += result.substring(pos); break }
-        final += result.substring(pos, start) + '<code class="dw-inline-code">' + result.substring(start + 1, end) + '</code>'
-        pos = end + 1
-      }
-      out += final + '\n'
-    }
-    if (inCode) out += '</code></pre>'
-    return out
-  }
-
-  // --- Parse weave text into turns ---
-
-  const SPEAKERS = ['human', 'assistant', 'tool', 'edit', 'read', 'search', 'write', 'hook', 'session', 'compaction', 'agent', 'task']
-
-  function isSpeakerLine(s: string): boolean {
-    if (s[0] !== '[') return false
-    const end = s.indexOf('] ')
-    if (end < 1) return false
-    return SPEAKERS.includes(s.substring(1, end))
-  }
-
-  function parseTurns(weave: Weave): Turn[] {
-    if (!weave.text) return []
-    const parts = weave.text.split('\n\n')
-    // Re-join parts that don't start with a speaker label onto the previous chunk
-    const chunks: string[] = []
-    for (const part of parts) {
-      if (isSpeakerLine(part) || chunks.length === 0) {
-        chunks.push(part)
-      } else {
-        chunks[chunks.length - 1] += '\n\n' + part
-      }
-    }
-    const turns: Turn[] = []
-    for (let i = 0; i < chunks.length; i++) {
-      const chunk = chunks[i].trim()
-      if (!chunk) continue
-      const end = chunk.indexOf('] ')
-      if (chunk[0] === '[' && end > 0) {
-        const speaker = chunk.substring(1, end)
-        const text = chunk.substring(end + 2)
-        // Look up full path for file-related turns
-        let fullPath: string | undefined
-        if (weave.paths && (speaker === 'edit' || speaker === 'read' || speaker === 'write' || speaker === 'search')) {
-          fullPath = weave.paths[text]
-        }
-        turns.push({
-          speaker,
-          text,
-          weaveId: weave.id,
-          index: i,
-          timestamp: weave.timestamp,
-          branch: weave.branch,
-          fullPath,
-        })
-      } else {
-        turns.push({
-          speaker: 'unknown',
-          text: chunk,
-          weaveId: weave.id,
-          index: i,
-          timestamp: weave.timestamp,
-          branch: weave.branch,
-        })
-      }
-    }
-    return turns
-  }
 
   // --- Selection ---
-
-  function turnKey(t: Turn): string {
-    return `${t.weaveId}:${t.index}`
-  }
 
   function toggle(turn: Turn) {
     const k = turnKey(turn)
@@ -478,16 +355,6 @@
     navigator.clipboard.writeText(text)
   }
 
-  // --- Time formatting ---
-
-  function fmtTime(ts: number): string {
-    const d = new Date(ts)
-    const mon = d.toLocaleString('en', { month: 'short' })
-    const day = d.getDate()
-    const hh = String(d.getHours()).padStart(2, '0')
-    const mm = String(d.getMinutes()).padStart(2, '0')
-    return `${mon} ${day} ${hh}:${mm}`
-  }
 
   // --- Fetch data ---
 
@@ -862,50 +729,7 @@
               {#each spacers as sz}
                 <div class="dw-time-spacer" style="height: {sz}px"></div>
               {/each}
-              {@const wturns = parseTurns(weave)}
-                <div class="dw-weave" data-ts={weave.timestamp}>
-                  <div class="dw-wmeta">
-                    <span class="dw-copyable" onclick={(e) => { e.stopPropagation(); navigator.clipboard.writeText(weave.id) }}>{weave.branch}</span>
-                    {#if showClusters && clusterLabel(weave.id)}
-                      <span class="dw-cluster">{clusterLabel(weave.id)}</span>
-                    {/if}
-                    <span>{fmtTime(weave.timestamp)}</span>
-                    <span>{weave.word_count}w {weave.turn_count}t</span>
-                  </div>
-                  {#each wturns as turn}
-                    {@const k = turnKey(turn)}
-                    <div
-                      class="dw-turn"
-                      class:selected={selectedTurns.has(k)}
-                      class:human={turn.speaker === 'human'}
-                      class:assistant={turn.speaker === 'assistant'}
-                      class:tool={turn.speaker === 'tool'}
-                      class:edit={turn.speaker === 'edit'}
-                      class:read={turn.speaker === 'read'}
-                      class:search={turn.speaker === 'search'}
-                      class:write={turn.speaker === 'write'}
-                      class:hook={turn.speaker === 'hook'}
-                      class:marker={turn.speaker === 'session' || turn.speaker === 'compaction' || turn.speaker === 'agent' || turn.speaker === 'task'}
-                      onclick={() => toggle(turn)}
-                      role="button"
-                      tabindex="0"
-                      onkeydown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); toggle(turn) }}}
-                    >
-                      <span class="dw-speaker">[{turn.speaker}]</span>
-                      {#if turn.speaker === 'assistant'}
-                        <span class="dw-text">{@html renderText(turn.text)}</span>
-                      {:else if turn.fullPath}
-                        <span
-                          class="dw-text dw-path"
-                          title={turn.fullPath}
-                          onclick={(e) => { e.stopPropagation(); navigator.clipboard.writeText(turn.fullPath!) }}
-                        >{turn.text}</span>
-                      {:else}
-                        <span class="dw-text">{turn.text}</span>
-                      {/if}
-                    </div>
-                  {/each}
-                </div>
+              <Weave {weave} clusterLabel={clusterLabel(weave.id)} {selectedTurns} onToggle={toggle} />
             {/each}
           </div>
         </div>
@@ -968,7 +792,6 @@
 
   .dw-stat { color: var(--text-on-dark-tertiary); font-size: var(--font-size-sm); }
 
-  .dw-cluster { color: var(--text-on-dark-tertiary); font-style: italic; }
 
   .dw-sel {
     margin-left: auto;
@@ -1107,133 +930,6 @@
   /* Weave stream */
   .dw-stream { padding: 2px 0; flex: 1; }
 
-  .dw-weave {
-    border-left: 3px solid var(--border-on-dark);
-    margin: 0 6px 1px 6px;
-    padding: 3px 6px;
-  }
-
-  .dw-wmeta {
-    display: flex;
-    gap: 8px;
-    color: var(--text-secondary);
-    font-size: var(--font-size-xs);
-    padding-bottom: 2px;
-    border-bottom: 1px solid var(--bg-secondary);
-    margin-bottom: 1px;
-  }
-  .dw-wmeta span:last-child { margin-left: auto; }
-  .dw-copyable { cursor: pointer; }
-  .dw-copyable:hover { color: var(--text-on-dark-secondary); }
-
-  /* Turn */
-  .dw-turn {
-    padding: 0px 3px;
-    cursor: pointer;
-    user-select: none;
-    overflow-wrap: break-word;
-    word-break: break-word;
-  }
-
-  .dw-turn:hover { background: var(--bg-dark-hover); }
-  .dw-turn.selected { background: var(--glyph-status-running-bg); }
-
-  .dw-speaker {
-    font-weight: 500;
-    font-size: var(--font-size-xs);
-    margin-right: 4px;
-  }
-
-  .dw-turn.human .dw-speaker { color: var(--accent-on-dark); }
-  .dw-turn.assistant .dw-speaker { color: var(--glyph-status-running-text); }
-  .dw-turn.tool .dw-speaker { color: var(--color-warning); }
-  .dw-turn.marker .dw-speaker { color: var(--color-scheduled); }
-
-  .dw-turn.tool {
-    background: var(--bg-almost-black);
-    border-left: 2px solid var(--color-warning);
-    padding-left: 4px;
-    margin: 1px 0;
-  }
-  .dw-turn.tool .dw-text {
-    color: var(--text-on-dark-secondary);
-    font-size: 8px;
-  }
-
-  .dw-turn.edit,
-  .dw-turn.read,
-  .dw-turn.search,
-  .dw-turn.write {
-    background: var(--bg-almost-black);
-    border-left: 2px solid #5b8dd9;
-    padding-left: 4px;
-    margin: 1px 0;
-  }
-  .dw-turn.edit .dw-speaker { color: #5b8dd9; }
-  .dw-turn.read .dw-speaker { color: #5b8dd9; }
-  .dw-turn.search .dw-speaker { color: #5b8dd9; }
-  .dw-turn.write .dw-speaker { color: #5b8dd9; }
-  .dw-turn.edit .dw-text,
-  .dw-turn.read .dw-text,
-  .dw-turn.search .dw-text,
-  .dw-turn.write .dw-text {
-    color: var(--text-on-dark-secondary);
-    font-size: 8px;
-  }
-  .dw-path {
-    cursor: pointer;
-    border-bottom: 1px dotted var(--border-on-dark);
-  }
-  .dw-path:hover {
-    color: var(--text-on-dark);
-    border-bottom-color: #5b8dd9;
-  }
-
-  .dw-turn.hook {
-    background: var(--bg-almost-black);
-    border-left: 2px solid #d94a4a;
-    padding-left: 4px;
-    margin: 1px 0;
-  }
-  .dw-turn.hook .dw-speaker { color: #d94a4a; }
-  .dw-turn.hook .dw-text {
-    color: var(--text-on-dark-secondary);
-    font-size: 8px;
-  }
-
-  .dw-text {
-    font-size: 9px;
-    line-height: 1.05;
-    white-space: pre-wrap;
-    overflow-wrap: break-word;
-    word-break: break-word;
-  }
-
-  .dw-turn.marker {
-    font-size: 10px;
-    color: var(--text-secondary);
-    font-style: italic;
-  }
-  .dw-turn.marker .dw-text { color: var(--text-secondary); font-size: 10px; }
-
-  :global(.dw-code) {
-    background: var(--bg-almost-black);
-    border: 1px solid var(--border-on-dark);
-    padding: 3px 6px;
-    margin: 2px 0;
-    font-size: 11px;
-    overflow-wrap: break-word;
-    word-break: break-word;
-    white-space: pre-wrap;
-  }
-  :global(.dw-code code) { color: var(--text-on-dark-secondary); }
-
-  :global(.dw-inline-code) {
-    background: var(--bg-almost-black);
-    padding: 0 3px;
-    color: var(--text-on-dark-secondary);
-    font-size: 11px;
-  }
 
   /* Inline session list (per-project, in header) */
   .dw-sessions-btn {
